Start train in CLI
