+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "start": "node src/index.js",
     "prep-train": "node src/prepare-train.js",
+    "start-train": "node src/start-train.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
